IMPROVED: added mobx safety checks
Note that `observableRequiresReaction` is still set to false: when enabling it,
lots of warnings are reported because we access observable properties
inside constructors.

Maybe this code could be improved: I'm not sure wrapping observable access
withing constructors using action() is a good idea since we are not
modifying any state, only reading it.