ContextMenu: switched to ContextMenu v2
Added detection of ability to copy/paste: we had to add
a new state to FileTable: rightClickFile which keeps track
of the file that's under the mouse when a right click happens.

The FileContextMenu now has access to MobX stores and is only
rendered when the context menu is opened.

Still need to do:
- check when delete should be enabled
- implement copy/paste/delete

Streched goal:
- add a new outline that indicates which elements the action will act on
- update copy paste menu+context menu to include the number of elements (ie: "Paste 4 elements")