Drag & Drop: added overlay
When attempting to drop elements in a filetable, only a little outline indicated
if the target was valid. This PR adds a bigger overlay with an icon that clearly
indicates the target's validity.

The Loader component was renamed to Overlay and made generic so it can be
reused for the drop target.