appState: add toogleSplitViewMode that adds a fileCache if empty (#325)
There was no cache added when enabling splitviewMode. This crashed
SideView which was expecting one.

This PR adds a new appState.toggleSplitViewMode method that toggles split view
and adds a new cache using default path if needed.

Also fixed to display an error on first load if an error occurred, instead of
showing the empty placeholder.