Statusbar add hidden files toggle (#332)
* StatusBar: replace paste button with show hidden files toggle

Only UI is ready: need to add hidden files filter now.

* FileTable: respect cache.showHiddenFiles & reload view when it has changed

TODO: IIRC when a file is selected & is renamed outside of react-explorer, it's
automatically re-selected on cache reload. We should not re-select it if it becomes hidden
and showHiddenFiles is false :)

* StatusBar: use showHiddenFiles when counting the number of files/folders

Also do not re-select a file that is hidden when refreshing file cache.

* Statusbar: fixed tests

Also removed native-ext-loader from webpack+package.json since it's not needed anymore.

* Statusbar: don't keep a local showHiddenFiles state, rely on filecache instead

This will make it easier to modify showHiddenFiles from another component,
for example a native menu, shortcut,...

* Electron: added native menu to toggle hidden files

Menu can be activated using mod + h accelerator.

This combo was previously binded to the debug history command: I removed it
this was not used anymore.