WIP: rewrite FileView using hooks & modular layout architecture
A new useLayout hook has been added and returns the layout for the fileview contents.

Lots of code will be moved from FileView into this layout: this will not only reduce
the size of FileView but also allow to have different layouts (eg. IconView,...) for
the FileView :)

Lots of events are not implemented yet: this will need more work and a lot of testing
but should worth the pain ;)