IconLayout: fixed wrong layout ref
Problem from last commit was that the layout created in useLayout is actually
updated after the child component has rendered. If useLayout is not called again,
the returned layout would still be the previous one.

Also: select the first element and not the first element of the first row
when pressing down for the first time.