FsZip: fixed reading entries when dir is not listed
Some zip have a specific entry for directories inside the archive: 'foo/'
but some zips don't have it, so we have a manually add 'foo/' if the archive
contains for example 'foo/bar'.

Also added new 'indirect' option on Fs that will be used when opening a file
needs first to uncompress (eg. zip archive) or transfer (eg. FTP) file before
it can be opened by the OS.

Last but not least, onFsChange was made optional since archive Fs
don't have it (at least not in the initial version).