Toolbar: set input size to "1"
On a text input, the size sets the min width of the element so we set it to the
minimum value so that the input can shrink a lot more.

Note that this only mitigates the problem: because of the right padding of 30px,
there is still a little bit of overflow when the window is tiny.